move the common build process runner to it's own file
